Fix namespaces after move
