+++ analyzers/src/SonarAnalyzer.CSharp/SymbolicExecution/Sonar/Analyzers/PublicMethodArgumentsShouldBeCheckedForNull.cs
@@ -26,8 +26,6 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using SonarAnalyzer.Extensions;
 using SonarAnalyzer.Helpers;
-using SonarAnalyzer.Rules.SymbolicExecution;
-using SonarAnalyzer.SymbolicExecution.Sonar;
 using SonarAnalyzer.SymbolicExecution.Sonar.Constraints;
 
 namespace SonarAnalyzer.SymbolicExecution.Sonar.Analyzers
