+++ analyzers/src/SonarAnalyzer.CSharp/SymbolicExecution/Sonar/ISymbolicExecutionAnalyzer.cs
@@ -21,9 +21,8 @@
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
-using SonarAnalyzer.SymbolicExecution.Sonar;
 
-namespace SonarAnalyzer.Rules.SymbolicExecution
+namespace SonarAnalyzer.SymbolicExecution.Sonar
 {
     internal interface ISymbolicExecutionAnalyzer
     {
