+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/SymbolicExecution/EmptyCollectionsShouldNotBeEnumeratedTest.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using SonarAnalyzer.Rules.CSharp;
+using SonarAnalyzer.SymbolicExecution.Sonar;
 using SonarAnalyzer.SymbolicExecution.Sonar.Analyzers;
 
 namespace SonarAnalyzer.UnitTest.Rules.SymbolicExecution
