+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/SymbolicExecution/InvalidCastToInterfaceTest.cs
@@ -19,6 +19,7 @@
  */
 
 using SonarAnalyzer.Rules.CSharp;
+using SonarAnalyzer.SymbolicExecution.Sonar;
 using SonarAnalyzer.SymbolicExecution.Sonar.Analyzers;
 
 namespace SonarAnalyzer.UnitTest.Rules.SymbolicExecution
