+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/SymbolicExecution/Roslyn/LocksReleasedAllPathsTest.cs
@@ -18,7 +18,6 @@
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 
-using SonarAnalyzer.Common;
 using Checks = SonarAnalyzer.SymbolicExecution.Roslyn.RuleChecks;
 using CS = SonarAnalyzer.Rules.CSharp;
 using VB = SonarAnalyzer.Rules.VisualBasic;
